requirements.txt
# LiveKit Agents Core
livekit-agents[deepgram,openai,cartesia,elevenlabs,silero]==1.0.23

# Additional LiveKit plugins
livekit-plugins-silero==1.0.23

# Data processing and Excel export
pandas>=2.0.0
openpyxl>=3.1.0

# Environment management
python-dotenv>=1.0.0

# Audio processing
soundfile>=0.12.0
pyaudio>=0.2.11

# Async HTTP
aiohttp>=3.8.0

# Logging
coloredlogs>=15.0.0

# Type hints
typing-extensions>=4.5.0 